% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-get_wpp2019.R
\name{get_wpp2019_ageSex}
\alias{get_wpp2019_ageSex}
\title{Get population by age and sex estimates and projections by medium variant
and other variants from the World Population Prospects 2019}
\usage{
get_wpp2019_ageSex()
}
\value{
A data.frame of population age and sex estimates and projections by
  medium variant and other variants from the World Population Prospects 2019
}
\description{
Get population by age and sex estimates and projections by medium variant
and other variants from the World Population Prospects 2019
}
\examples{
get_wpp2019_ageSex()

}
