% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-get_wpp2019.R
\name{get_wpp2019_total}
\alias{get_wpp2019_total}
\title{Get all variants total population estimates and projections from the World
Population Prospects 2019}
\usage{
get_wpp2019_total()
}
\value{
A data.frame of all variants total population estimates and
  projections from the World Population Prospects 2019
}
\description{
Get all variants total population estimates and projections from the World
Population Prospects 2019
}
\examples{
get_wpp2019_total()

}
