% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populi.R
\docType{package}
\name{populi}
\alias{populi}
\alias{populi-package}
\title{The Population Division of the Department of Economic and Social Affairs of
the United Nations Secretariat prepares regular revisions of the World
Population Prospects - the official United Nations population estimates and
projections. This package provides utilities and tools to extract, process,
and structure these population datasets for use in R.}
\description{
The Population Division of the Department of Economic and Social
    Affairs of the United Nations Secretariat prepares regular revisions of the
    World Population Prospects - the official United Nations population
    estimates and projections. This package provides utilities and tools to 
    extract, process, and structure these population datasets for use in R.
}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
